services/ingestion/requirements.txt
fastapi
uvicorn
python-multipart
minio
confluent-kafka
python-dotenv
psycopg2-binary
